pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "md2docx"
version = "1.0.0"
description = "一個具備圖形化介面的桌面工具，可將 Markdown (.md) 檔案精確轉換為 Microsoft Word (.docx) 檔案"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "md2docx Team"}
]
keywords = ["markdown", "word", "docx", "converter", "gui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "customtkinter>=5.2.1",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "pyinstaller>=6.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/md2docx"
"Bug Reports" = "https://github.com/yourusername/md2docx/issues"
"Source" = "https://github.com/yourusername/md2docx"

[project.scripts]
md2docx = "md_to_word_converter:main"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=. --cov-report=html --cov-report=term-missing"
testpaths = [
    "test_converter.py",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]
